% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example Data}
\format{
Originally CSV
}
\source{
Flickr API collected by author
}
\usage{
exampledata
}
\description{
Anonymized example data from Flickr API
}
\keyword{datasets}
