% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpticsXtractDBSCAN.R
\name{OpticsXtractDBSCAN}
\alias{OpticsXtractDBSCAN}
\title{Extracts DBScan Optics from point data}
\usage{
OpticsXtractDBSCAN(filename, x, y, eps, minPts, eps_cl)
}
\arguments{
\item{filename}{a csv containing x,y coordinates}

\item{x}{column number with x coordinates}

\item{y}{column number with y coordinates}

\item{eps}{EPS parameter for clustering}

\item{minPts}{minPts parameter for clustering}

\item{eps_cl}{parameter for cluster extraction}
}
\value{
a dataframe with the csv content and the cluster ID added as column
}
\description{
Extracts DBScan Optics from point data
}
